+++ generators/new-wp-plugin/readme.md
