+++ generators/new-wp-project/readme.md
@@ -1,0 +1,19 @@
+# yo moxie-lean:new-wp
+
+> Generates the config we need for a WordPress project.
+
+To run this command just type on the terminal: 
+
+```bash
+yo moxie-lean:new-wp
+```
+
+You need to set-up the remote repo first on GitHub or Bitbucket.
+
+This generator creates: 
+
+- A composer.json file
+- A default .gitignore file
+- A .travis.yml file to set up CI.
+
+It then initialises Git and pushes to your remote repo.
