+++ generators/new-wp-project/templates/composer.json
