+++ generators/new-wp-project/templates/gitignore
