+++ generators/new-wp-project/templates/travis.yml
