+++ generators/new-wp/readme.md
@@ -1,19 +1,11 @@
-# yo moxie-lean:new-wp
+# yo moxie-lean:new-plugin
 
-> Generates the config we need for a WordPress project.
+> Generates a new WordPress plugin.
 
 To run this command just type on the terminal: 
 
 ```bash
-yo moxie-lean:new-wp
+yo moxie-lean:new-plugin
 ```
 
-You need to set-up the remote repo first on GitHub or Bitbucket.
-
-This generator creates: 
-
-- A composer.json file
-- A default .gitignore file
-- A .travis.yml file to set up CI.
-
-It then initialises Git and pushes to your remote repo.
+This generator creates a new plugin and updates the namespace and constant names.
